details config
Modified the dashboard component to store the _id property of the exoplanet object in the URL when the <Link> component is clicked.
Added the useEffect and useState hooks to manage the selected exoplanet ID in the dashboard component.
Updated the onClick event handler of the <Link> component to update the selected exoplanet ID using setSelectedExoplanetId.
Passed the selected exoplanet ID as a query parameter to the /details page using template literals.
Modified the getServerSideProps function in the dashboard component to retrieve the exoplanet ID from the query parameter.
Imported the useEffect and useState hooks in the details component.
Modified the details component to retrieve the exoplanet ID from the query parameter as a prop.
Updated the getServerSideProps function in the details component to retrieve the exoplanet ID from the query parameter and pass it as a prop.