adds three new API endpoints related to bookmarking functionality:
checkIfBookmarked API endpoint:

This endpoint checks if a card is bookmarked by making a request to the server.
It takes the exoplanetId and userEmail as parameters.
The server checks if the card is bookmarked for the specified user and returns a response indicating whether it is bookmarked or not.
This endpoint helps determine whether to display the "Save" or "Saved" button on the card.
applyBookmark API endpoint:

This endpoint applies a bookmark to a card for a specific user.
It takes the exoplanetId and userEmail as parameters.
The server updates the user's document in the database, setting the planetFav property with the provided exoplanetId.
This endpoint is called when the user clicks the "Save" button on the card, and it saves the card as a bookmark for the user.
removeBookmark API endpoint:

This endpoint removes a bookmark from a card for a specific user.
It takes the exoplanetId and userEmail as parameters.
The server updates the user's document in the database, setting the planetFav property to null or removing it altogether.
This endpoint is called when the user clicks the "Saved" button on the card, and it removes the card from the user's bookmarks.