Change data type for values in planetFav property of users documents
This commit changes the data type for the values in the planetFav property of the users documents from an _id reference to the name of the planets. The planet names themselves are unique and specific, acting as identifiers. Therefore, there are no concerns about data integrity. The update ensures that the planet names are stored directly in the planetFav array, providing a more descriptive representation of the favorite planets for each user.